+++ sonar-groovy-plugin/src/main/java/org/sonar/plugins/groovy/jacoco/JaCoCoSensor.java
@@ -20,16 +20,23 @@
 package org.sonar.plugins.groovy.jacoco;
 
 import java.io.File;
+import java.nio.file.Files;
 import java.nio.file.Path;
+import java.util.Arrays;
 import org.sonar.api.batch.sensor.Sensor;
 import org.sonar.api.batch.sensor.SensorContext;
 import org.sonar.api.batch.sensor.SensorDescriptor;
 import org.sonar.api.config.Settings;
+import org.sonar.api.notifications.AnalysisWarnings;
 import org.sonar.api.scan.filesystem.PathResolver;
+import org.sonar.api.utils.log.Logger;
+import org.sonar.api.utils.log.Loggers;
 import org.sonar.plugins.groovy.foundation.Groovy;
 import org.sonar.plugins.groovy.foundation.GroovyFileSystem;
 
 public class JaCoCoSensor implements Sensor {
+
+  private static final Logger LOG = Loggers.get(JaCoCoSensor.class.getName());
 
   public static final String JACOCO_OVERALL = "jacoco-overall.exec";
 
@@ -37,16 +44,23 @@ public class JaCoCoSensor implements Sensor {
   private final GroovyFileSystem fileSystem;
   private final PathResolver pathResolver;
   private final Settings settings;
+  private final AnalysisWarnings analysisWarnings;
+  static final String JACOCO_XML_PROPERTY = "sonar.coverage.jacoco.xmlReportPaths";
+  private static final String[] JACOCO_XML_DEFAULT_PATHS = {
+    "target/site/jacoco/jacoco.xml", "build/reports/jacoco/test/jacocoTestReport.xml"
+  };
 
   public JaCoCoSensor(
       JaCoCoConfiguration configuration,
       GroovyFileSystem fileSystem,
       PathResolver pathResolver,
-      Settings settings) {
+      Settings settings,
+      AnalysisWarnings analysisWarnings) {
     this.configuration = configuration;
     this.fileSystem = fileSystem;
     this.pathResolver = pathResolver;
     this.settings = settings;
+    this.analysisWarnings = analysisWarnings;
   }
 
   @Override
@@ -56,22 +70,65 @@ public class JaCoCoSensor implements Sensor {
 
   @Override
   public void execute(SensorContext context) {
+    boolean hasXmlReport = hasXmlReport(context);
     File baseDir = fileSystem.baseDir();
     File reportUTs = pathResolver.relativeFile(baseDir, configuration.getReportPath());
     File reportITs = pathResolver.relativeFile(baseDir, configuration.getItReportPath());
-    if (shouldExecuteOnProject()) {
+
+    if (reportUTs.isFile()) {
+      warnAboutDeprecatedProperty(hasXmlReport, JaCoCoConfiguration.REPORT_PATH_PROPERTY);
+    }
+    if (reportITs.isFile()) {
+      warnAboutDeprecatedProperty(hasXmlReport, JaCoCoConfiguration.IT_REPORT_PATH_PROPERTY);
+    }
+    if (hasXmlReport) {
+      LOG.debug(
+          "JaCoCo XML report found, skipping processing of binary JaCoCo exec report.",
+          JACOCO_XML_PROPERTY);
+      return;
+    }
+
+    if (shouldExecuteOnProject(reportUTs.isFile(), reportITs.isFile())) {
       Path reportOverall = context.fileSystem().workDir().toPath().resolve(JACOCO_OVERALL);
       JaCoCoReportMerger.mergeReports(reportOverall, reportUTs, reportITs);
       new JaCoCoAnalyzer(fileSystem, settings, reportOverall).analyse(context);
     }
   }
 
+  private void warnAboutDeprecatedProperty(boolean hasXmlReport, String deprecatedProperty) {
+    if (!hasXmlReport) {
+      addAnalysisWarning(
+          "Property '%s' is deprecated (JaCoCo binary format). '%s' should be used instead (JaCoCo XML format)."
+              + " Please check that the JaCoCo plugin is installed on your SonarQube Instance.",
+          deprecatedProperty, JACOCO_XML_PROPERTY);
+    } else if (settings.hasKey(deprecatedProperty)) {
+      // only log for those properties which were set explicitly
+      LOG.info(
+          "Both '{}' and '{}' were set. '{}' is deprecated therefore, only '{}' will be taken into account."
+              + " Please check that the JaCoCo plugin is installed on your SonarQube Instance.",
+          deprecatedProperty,
+          JACOCO_XML_PROPERTY,
+          deprecatedProperty,
+          JACOCO_XML_PROPERTY);
+    }
+  }
+
+  private static boolean hasXmlReport(SensorContext context) {
+    return context.config().hasKey(JACOCO_XML_PROPERTY)
+        || Arrays.stream(JACOCO_XML_DEFAULT_PATHS)
+            .map(path -> context.fileSystem().baseDir().toPath().resolve(path))
+            .anyMatch(Files::isRegularFile);
+  }
+
+  private void addAnalysisWarning(String format, Object... args) {
+    String msg = String.format(format, args);
+    LOG.warn(msg);
+    analysisWarnings.addUnique(msg);
+  }
+
   // VisibleForTesting
-  boolean shouldExecuteOnProject() {
-    File baseDir = fileSystem.baseDir();
-    File reportUTs = pathResolver.relativeFile(baseDir, configuration.getReportPath());
-    File reportITs = pathResolver.relativeFile(baseDir, configuration.getItReportPath());
-    boolean foundOneReport = reportUTs.isFile() || reportITs.isFile();
+  boolean shouldExecuteOnProject(boolean hasUT, boolean hasIT) {
+    boolean foundOneReport = hasUT || hasIT;
     boolean shouldExecute = configuration.shouldExecuteOnProject(foundOneReport);
     if (!foundOneReport && shouldExecute) {
       JaCoCoExtensions.logger().info("JaCoCoSensor: No JaCoCo report found.");